requirements.txt
defusedxml==0.4.1
python-slugify==1.2.0
